feat: Add Login with github, facebook and twitter
So that a user can login to their store and come back to their order, I want to create the ability to login via login API's to enable the user to do this. I have setup 3 buttons on the page and have use firebase to handle the login process. In this commit the user can logs in and the response authentification details are printed to the console while in development.